App/backend/requirements.txt
fastapi==0.111.0
huggingface-hub==0.23.0
numpy==1.26.4
torch==2.3.0
transformers==4.41.0
uvicorn==0.29.0
weaviate-client==4.6.3
python-dotenv==1.0.1
